pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "AI Agent utilizing LangChain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.24",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "wikipedia>=1.4.0",
]
